Bug fixes, small changes

+++ polls/views.py
@@ -1,15 +1,11 @@
-import time
-
 from django.contrib.auth.models import User
 from django.http import HttpResponse
 
-from django.http.response import JsonResponse
 from django.shortcuts import redirect
 from rest_framework.parsers import JSONParser
 from rest_framework import status
 
 from polls.models import User, Country
-from polls.serializers import UserSerializer, CountrySerializer
 from rest_framework.decorators import api_view
 from mongoengine import *
 
@@ -19,7 +15,6 @@ from polls.services.game_service import GameService
 from polls.services.system_service import SystemService
 from polls.services.user_service import UserService
 from polls.services.view_service import CountryViewService, NewsViewService
-from polls.view_models.basic_statistic import BasicStatisticView
 
 connect('TestDb')
 
@@ -30,10 +25,11 @@ def index(request):
 def login(request,username,password):
     if request.method == 'POST':
         try:
+            user = User.objects(username=username).first()
             result_bool = UserService().login(username,password)
         except Exception as error:
             return HttpResponse(error, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-        return HttpResponse({}, status=status.HTTP_202_ACCEPTED) if result_bool else HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
+        return HttpResponse(user._id, status=status.HTTP_202_ACCEPTED) if result_bool else HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
     else:
         return HttpResponse({}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
 
@@ -54,9 +50,11 @@ def register(request):
         try:
             request_data = JSONParser().parse(request)
             result_bool = UserService().register_new_user(request_data['username'],request_data['password'],request_data['email'],request_data['country_name'],request_data['link_on_flag'])
+            if result_bool:
+                UserService().login(request_data['username'],request_data['password'])
         except Exception as error:
             return HttpResponse(error, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-        return HttpResponse({}, status=status.HTTP_201_CREATED) if result_bool else HttpResponse({},status=status.HTTP_400_BAD_REQUEST)
+        return HttpResponse(User.objects(username=request_data['username']).first()._id, status=status.HTTP_201_CREATED) if result_bool else HttpResponse({},status=status.HTTP_400_BAD_REQUEST)
     else:
         return HttpResponse({}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
 
@@ -64,6 +62,8 @@ def register(request):
 def delete_account(request,user_id,password):
     if request.method == 'DELETE':
         try:
+            if not User.objects(_id=user_id).first().isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
             result_bool = UserService().delete_user_account(user_id,password)
         except Exception as error:
             return HttpResponse(error, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
@@ -75,6 +75,8 @@ def delete_account(request,user_id,password):
 def change_user_data(request,user_id):
     if request.method == 'PUT':
         try:
+            if not User.objects(_id=user_id).first().isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
             request_data = JSONParser().parse(request)
             result_bool = UserService().change_user_data(user_id,request_data['username'],request_data['password'],request_data['country_name'],request_data['link_on_flag'])
         except Exception as error:
@@ -88,6 +90,8 @@ def redirect_feedback(request,user_id):
     if request.method == 'POST':
         try:
             user = User.objects(_id=user_id).first()
+            if not user.isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
             request_data = JSONParser().parse(request)
             if user is not None:
                 SystemService().get_feedback(request_data['username'],request_data['email'],request_data['rating'],request_data['msg'])
@@ -99,22 +103,34 @@ def redirect_feedback(request,user_id):
         return HttpResponse({}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
 
 @api_view(['GET'])
-def get_all(request,user_id):
+def get_view(request,user_id,name_view):
     if request.method == 'GET':
         try:
-            obj = {
-                'basic statistic':CountryViewService().get_basic_statistic(user_id),
-                'budget':CountryViewService().get_budget(user_id),
-                'technologies':CountryViewService().get_technologies(user_id),
-                'industry':CountryViewService().get_industry(user_id),
-                'warehouses':CountryViewService().get_warehouses(user_id),
-                'adopted politics laws':CountryViewService().get_politics_laws(user_id),
-                'population':CountryViewService().get_population(user_id),
-                'trade':CountryViewService().get_trade(user_id),
-                'army':CountryViewService().get_army(user_id),
-                'news':NewsViewService().get_news()
-            }
-            return HttpResponse(json.dumps(obj, default=lambda x: x.__dict__), status=status.HTTP_200_OK)
+            if not User.objects(_id=user_id).first().isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
+            view_obj = None
+            if name_view == 'BasicStatistic':
+                view_obj = CountryViewService().get_basic_statistic(user_id)
+            elif name_view == 'Budget':
+                view_obj = CountryViewService().get_budget(user_id)
+            elif name_view == 'Technologies':
+                view_obj = CountryViewService().get_technologies(user_id)
+            elif name_view == 'Industry':
+                view_obj = CountryViewService().get_industry(user_id)
+            elif name_view == 'Warehouses':
+                view_obj = CountryViewService().get_warehouses(user_id)
+            elif name_view == 'Politics':
+                view_obj = CountryViewService().get_politics_laws(user_id)
+            elif name_view == 'Population':
+                view_obj = CountryViewService().get_population(user_id)
+            elif name_view == 'Trade':
+                view_obj = CountryViewService().get_trade(user_id)
+            elif name_view == 'Army':
+                view_obj = CountryViewService().get_army(user_id)
+            elif name_view == 'News':
+                view_obj = NewsViewService().get_news()
+
+            return HttpResponse(json.dumps(view_obj, default=lambda x: x.__dict__), status=status.HTTP_200_OK)
         except Exception as error:
             return HttpResponse(error, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
     else:
@@ -125,6 +141,8 @@ def change_taxes(request,user_id):
     if request.method == 'PUT':
         try:
             user = User.objects(_id=user_id).first()
+            if not user.isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
             country = Country.objects(_id=user.country._id).first()
             request_data = JSONParser().parse(request)
             GameService().set_taxes(country.name,request_data['name_tax'],request_data['new_value'])
@@ -139,6 +157,8 @@ def upgrade_technology(request,user_id):
     if request.method == 'PUT':
         try:
             user = User.objects(_id=user_id).first()
+            if not user.isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
             country = Country.objects(_id=user.country._id).first()
             request_data = JSONParser().parse(request)
             GameService().upgrade_technology(country.name, request_data['name_technology'])
@@ -153,6 +173,8 @@ def build_industry(request,user_id):
     if request.method == 'PUT':
         try:
             user = User.objects(_id=user_id).first()
+            if not user.isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
             country = Country.objects(_id=user.country._id).first()
             request_data = JSONParser().parse(request)
             GameService().build_industry(country.name,request_data['name_building'])
@@ -167,6 +189,8 @@ def remove_industry(request,user_id):
     if request.method == 'PUT':
         try:
             user = User.objects(_id=user_id).first()
+            if not user.isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
             country = Country.objects(_id=user.country._id).first()
             request_data = JSONParser().parse(request)
             GameService().remove_industry(country.name, request_data['name_building'])
@@ -181,6 +205,8 @@ def upgrade_warehouse(request,user_id):
     if request.method == 'PUT':
         try:
             user = User.objects(_id=user_id).first()
+            if not user.isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
             country = Country.objects(_id=user.country._id).first()
             request_data = JSONParser().parse(request)
             GameService().upgrade_warehouse(country.name, request_data['name_warehouse'])
@@ -195,10 +221,12 @@ def set_law(request,user_id):
     if request.method == 'PUT':
         try:
             user = User.objects(_id=user_id).first()
+            if not user.isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
             country = Country.objects(_id=user.country._id).first()
             request_data = JSONParser().parse(request)
-            names_political_laws = ['Immigration','Isolation','Free housing','Free education']
-            if remove_industry.name_law in names_political_laws:
+            names_political_laws = ['Isolation','Free medicine','Free housing','Free education']
+            if request_data['name_law'] in names_political_laws:
                 GameService().set_politics_law(country.name, request_data['name_law'])
             else:
                 GameService().set_conscript_law(country.name, request_data['name_law'])
@@ -213,6 +241,8 @@ def cancel_law(request,user_id):
     if request.method == 'PUT':
         try:
             user = User.objects(_id=user_id).first()
+            if not user.isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
             country = Country.objects(_id=user.country._id).first()
             request_data = JSONParser().parse(request)
             GameService().cancel_politics_law(country.name, request_data['name_law'])
@@ -227,6 +257,8 @@ def buy_goods(request,user_id):
     if request.method == 'POST':
         try:
             user = User.objects(_id=user_id).first()
+            if not user.isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
             country = Country.objects(_id=user.country._id).first()
             request_data = JSONParser().parse(request)
             GameService().buy_goods(country.name, request.request_data['name_goods'],request_data['number'])
@@ -241,6 +273,8 @@ def sell_goods(request,user_id):
     if request.method == 'POST':
         try:
             user = User.objects(_id=user_id).first()
+            if not user.isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
             country = Country.objects(_id=user.country._id).first()
             request_data = JSONParser().parse(request)
             GameService().sell_goods(country.name, request.request_data['name_goods'],request_data['number'])
@@ -256,6 +290,8 @@ def edit_army(request,user_id):
     if request.method == 'PUT':
         try:
             user = User.objects(_id=user_id).first()
+            if not user.isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
             country = Country.objects(_id=user.country._id).first()
             request_data = JSONParser().parse(request)
             GameService().edit_army(country.name,request_data['name_unit'],request_data['new_number'])
@@ -270,6 +306,8 @@ def calculate_war(request,user_id,defending_player_name):
     if request.method == 'POST':
         try:
             user_1,user_2 = User.objects(_id=user_id).first(),User.objects(username=defending_player_name).first()
+            if not user_1.isAuth:
+                return HttpResponse({}, status=status.HTTP_401_UNAUTHORIZED)
             country_1,country_2 = Country.objects(_id=user_1.country._id).first(),Country.objects(_id=user_2.country._id).first()
             GameService().calculate_war(country_1.name,country_2.username)
             return HttpResponse({}, status=status.HTTP_200_OK)
@@ -277,30 +315,5 @@ def calculate_war(request,user_id,defending_player_name):
             return HttpResponse(error, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
     else:
         return HttpResponse({}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
-
-
-
-
-
-
-
-@api_view(['GET', 'POST', 'DELETE'])
-def tutorial_list(request):
-    if request.method == 'GET':
-        #print(UserService().register_new_user('ar3', '123', 'test@gmail.com', 'c4', 'img'))
-
-        news = Country.objects(id='5f1dc17211fc3660885d8d24').only('name')
-        serializer = CountrySerializer(news,many=True)
-        bs = BasicStatisticView()
-
-        return JsonResponse(serializer.data,safe=False)
-
-    if request.method == 'POST':
-        tutorial_data = JSONParser().parse(request)
-        tutorial_serializer = UserSerializer(data=tutorial_data)
-        if tutorial_serializer.is_valid():
-            tutorial_serializer.save()
-            return JsonResponse(tutorial_serializer.data, status=status.HTTP_201_CREATED)
-        return JsonResponse(tutorial_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
